Share code between chimeraX and pymol projections
Shared helper functions moved to new file project_structure_utils.py
