+++ docs/project_structure_utils.md
@@ -1,0 +1,1 @@
+::: key_interactions_finder.project_structure_utils
